Move Header inside Components directory
