+++ client/app.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import * as ReactDom from 'react-dom'
-import { Header } from './header/Header';
+import { Header } from './Components/Header/Header';
 import { MainRouter } from './router';
 
 export class App extends React.Component<{}, {}> {
