+++ client/router.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react'
-import { Router, Route } from 'react-router'
-import { Header } from './header/Header';
+import { Header } from './Components/Header/Header';
 import { BrowserRouter } from 'react-router-dom';
 
 export class MainRouter extends React.Component<{}, {}> {
